Implement TCE download with dependency installation...
The build script had a list of TCE extensions to install or mirror, which
included a hand-maintained list of all the dependencies of the extensions.

That means a maintenance headache as we change them, especially as new
extensions are added or when TCL up-rev their tools.

This implements a script to mirror a set of TCE packages, and their
dependencies, allowing us to delegate that work to upstream.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>