Use 7z to extract ISO image without root required
The earlier ISO build process required root access to be able to
loopback-mount the TCL ISO image.  This was only done to enable access to the
embedded files, so that we could repack and rebuild the image later.

The `7z` utility (aka [`p7zip` or `p7zip-full`][p7zip]) can also extract an
ISO image, by treating it as a regular archive like a tar or zip.  This avoids
the need to be root, to mount images, and other messing about at the cost of
an extra tool dependency.

It also eliminates the only phase of rebuilding the ISO and TCL core that
actually *required* real root access - making it possible to rebuild the image
using `fakeroot`.

Since [p7zip is widely available, this doesn't seem like an extreme
burden][p7zip].

[p7zip]: http://pkgs.org/search/?keyword=p7zip

Signed-off-by: Daniel Pittman <daniel@rimspace.net>